requirements.txt
autopep8==1.5.7
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.5.30
charset-normalizer==2.0.5
greenlet==1.1.2
idna==3.2
lxml==4.6.5
numpy==1.21.5
pandas==1.3.5
pyaml==21.10.1
pycodestyle==2.7.0
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.26.0
six==1.16.0
soupsieve==2.3.1
SQLAlchemy==1.4.29
toml==0.10.2
urllib3==1.26.6
